+++ tests/engine/cloud/test_cloud_flow_runner.py
@@ -1,5 +1,7 @@
-import uuid
+import datetime
+import pendulum
 import time
+import uuid
 from datetime import timedelta
 from unittest.mock import MagicMock
 
@@ -17,6 +19,7 @@ from prefect.engine.state import (
     Pending,
     Retrying,
     Running,
+    Scheduled,
     Skipped,
     Success,
     TimedOut,
@@ -220,7 +223,6 @@ def test_flow_runner_loads_parameters_from_cloud(monkeypatch):
 
 
 def test_flow_runner_loads_context_from_cloud(monkeypatch):
-
     flow = prefect.Flow(name="test")
     get_flow_run_info = MagicMock(return_value=MagicMock(context={"a": 1}))
     set_flow_run_state = MagicMock()
@@ -235,6 +237,46 @@ def test_flow_runner_loads_context_from_cloud(monkeypatch):
     )
 
     assert res.context["a"] == 1
+
+
+def test_flow_runner_puts_scheduled_start_time_in_context(monkeypatch):
+    flow = prefect.Flow(name="test")
+    date = pendulum.parse("19860920")
+    get_flow_run_info = MagicMock(
+        return_value=MagicMock(context={}, scheduled_start_time=date)
+    )
+    set_flow_run_state = MagicMock()
+    client = MagicMock(
+        get_flow_run_info=get_flow_run_info, set_flow_run_state=set_flow_run_state
+    )
+    monkeypatch.setattr(
+        "prefect.engine.cloud.flow_runner.Client", MagicMock(return_value=client)
+    )
+    res = CloudFlowRunner(flow=flow).initialize_run(
+        state=None, task_states={}, context={}, task_contexts={}, parameters={}
+    )
+
+    assert "scheduled_start_time" in res.context
+    assert isinstance(res.context["scheduled_start_time"], datetime.datetime)
+    assert res.context["scheduled_start_time"].strftime("%Y-%m-%d") == "1986-09-20"
+
+
+def test_flow_runner_prioritizes_user_context_over_default_context(monkeypatch):
+    flow = prefect.Flow(name="test")
+    get_flow_run_info = MagicMock(return_value=MagicMock(context={"today": "is a day"}))
+    set_flow_run_state = MagicMock()
+    client = MagicMock(
+        get_flow_run_info=get_flow_run_info, set_flow_run_state=set_flow_run_state
+    )
+    monkeypatch.setattr(
+        "prefect.engine.cloud.flow_runner.Client", MagicMock(return_value=client)
+    )
+    res = CloudFlowRunner(flow=flow).initialize_run(
+        state=None, task_states={}, context={}, task_contexts={}, parameters={}
+    )
+
+    assert "today" in res.context
+    assert res.context["today"] == "is a day"
 
 
 def test_client_is_always_called_even_during_failures(client):